make a component to reuse functions for breadcrumbs
