+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/components/contests/contest-breadcrumb/ContestBreadcrumb.module.scss
@@ -1,0 +1,13 @@
+﻿@use 'src/styles/colors';
+
+.breadcrumbBtn {
+  color: colors.$light-blue-color;
+
+  &:hover {
+    color: colors.$text-dark-gray-color;
+  }
+}
+
+.breadcrumbBtnLast {
+  color: colors.$text-dark-gray-color;
+}
