+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/pages/contest/ContestDetailsPage.module.scss
@@ -31,14 +31,6 @@
   align-self: center;
 }
 
-.breadcrumbBtn {
-  color: colors.$light-blue-color;
-
-  &:hover {
-    color: colors.$text-dark-gray-color;
-  }
-}
-
 .buttonsContainer {
   column-gap: 1%;
   display: flex;
